+++ public/index.html
@@ -26,5 +26,5 @@
 </footer>
 
 </body>
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </html>
