+++ src/view/most-commented-films.js
@@ -1,0 +1,10 @@
+export const createMostCommentedFilmsTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Most commented</h2>
+
+      <div class="films-list__container">
+      </div>
+    </section>`
+  );
+};
